fixed dumb ass texture index problem on windows

+++ assets/shaders/test.vert
@@ -2,7 +2,7 @@
 layout (location = 0) in vec3 a_position;
 layout (location = 1) in vec3 a_normal;
 layout (location = 2) in vec2 a_texture_coord;
-layout (location = 3) in int a_texture_index;
+layout (location = 3) in float a_texture_index;
 
 out vec2 texture_coord;
 out float texture_index;
